+++ SpriteLib3.0-v2.0/SpriteLib3.0-v2.0/Scene.h
@@ -28,7 +28,7 @@ public:
 	virtual void GUI();
 
 	void AdjustScrollOffset();
-	void CreateDecoration(std::string name, int sprX, int sprY, int posX, int posY, int posZ);
+	void CreateDecoration(std::string name, float sprX, float sprY, float posX, float posY, float posZ);
 	void CreateCameraEntity(bool mainCamera, float windowWidth, float windowHeight, float left, float right, float bottom, float top, 
 								float zNear, float zFar, float aspectRatio, bool vertScroll=false, bool horizScroll=false);
 	void CreateBoxEntity(std::string fileName, int spriteX, int spriteY, int vecX, int vecY, int rotDeg = 0, int vecZ = 1.f, bool isDynamic = false, float shrinkXValue = 0, float shrinkYValue = 0);
